requirements.txt
certifi==2018.11.29
chardet==3.0.4
coverage==4.5.2
Django==2.1.3
django-extensions==2.1.4
djangorestframework==3.9.0
docopt==0.6.2
idna==2.8
nose==1.3.7
pytz==2018.7
requests==2.21.0
six==1.12.0
urllib3==1.24.1
